# Add method and benchmarks for pooling metric options