--- a/instrumentation/net/http/otelhttp/internal/semconv/env.go
+++ b/instrumentation/net/http/otelhttp/internal/semconv/env.go
@@ -103,30 +103,13 @@ type MetricData struct {
 	ElapsedTime float64
 }
 
-func (s HTTPServer) RecordMetrics(ctx context.Context, md ServerMetricData) {
-	if s.requestBytesCounter == nil || s.responseBytesCounter == nil || s.serverLatencyMeasure == nil {
-		// This will happen if an HTTPServer{} is used instead of NewHTTPServer.
-		return
-	}
-
-	attributes := oldHTTPServer{}.MetricAttributes(md.ServerName, md.Req, md.StatusCode, md.AdditionalAttributes)
-	o := metric.WithAttributeSet(attribute.NewSet(attributes...))
-	addOpts := []metric.AddOption{o}
-	s.requestBytesCounter.Add(ctx, md.RequestSize, addOpts...)
-	s.responseBytesCounter.Add(ctx, md.ResponseSize, addOpts...)
-	s.serverLatencyMeasure.Record(ctx, md.ElapsedTime, o)
-
-	// TODO: Duplicate Metrics
-}
-
 var metricAddOptionPool = &sync.Pool{
 	New: func() interface{} {
 		return &[]metric.AddOption{}
 	},
 }
 
-// RecordMetricsPooled is an implementation of RecordMetrics() where []metric.AddOption is pulled from a pool
-func (s HTTPServer) RecordMetricsPooled(ctx context.Context, md ServerMetricData) {
+func (s HTTPServer) RecordMetrics(ctx context.Context, md ServerMetricData) {
 	if s.requestBytesCounter == nil || s.responseBytesCounter == nil || s.serverLatencyMeasure == nil {
 		// This will happen if an HTTPServer{} is used instead of NewHTTPServer.
 		return
@@ -141,6 +124,8 @@ func (s HTTPServer) RecordMetricsPooled(ctx context.Context, md ServerMetricData
 	s.serverLatencyMeasure.Record(ctx, md.ElapsedTime, o)
 	*addOpts = (*addOpts)[:0]
 	metricAddOptionPool.Put(addOpts)
+
+	// TODO: Duplicate Metrics
 }
 
 func NewHTTPServer(meter metric.Meter) HTTPServer {