--- a/instrumentation/net/http/otelhttp/internal/semconv/env_test.go
+++ b/instrumentation/net/http/otelhttp/internal/semconv/env_test.go
@@ -153,18 +153,20 @@ func BenchmarkRecordMetrics(b *testing.B) {
 	}
 	b.ReportAllocs()
 	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		for _, tt := range testCases {
+	for _, tt := range testCases {
+		t.Run(tt.name, func(t *testing.T) {
 			req, _ := http.NewRequest("GET", "http://example.com", nil)
 
 			_ = tt.server.RequestTraceAttrs("stuff", req)
 			_ = tt.server.ResponseTraceAttrs(ResponseTelemetry{StatusCode: 200})
-			tt.server.RecordMetrics(context.Background(), ServerMetricData{
-				ServerName: "stuff",
-				MetricAttributes: MetricAttributes{
-					Req: req,
-				},
-			})
-		}
+			for i := 0; i < b.N; i++ {
+				tt.server.RecordMetrics(context.Background(), ServerMetricData{
+					ServerName: "stuff",
+					MetricAttributes: MetricAttributes{
+						Req: req,
+					},
+				})
+			}
+		})
 	}
 }