@vordimous 2024-12-05
@dmathieu I implemented the feedback and followed the `BenchmarkCoreWrite()` to have each case benchmarked. 

I also created the spelling fix PR: https://github.com/open-telemetry/opentelemetry-go-contrib/pull/6407